libvpd: Corrects data type of variables
In the below snippet the result of comparison between 'string::npos'
(size_t) and 'end' (unsigned int) variable was always a false statement
catched by coverity tool and same is the case with comparison between
'string::npos' and 'z' (unsigned int) variable.

Snippet:
 ...

   if ( string::npos == (end = s1.find("*", beg))) {
	//No more stars - base case
        if (string::npos != (z = s2.find(s1.substr(beg, end), 0))) {
		return true;
	 ...

This patch fixes the issue by changing the data type of 'end' and 'z'
to size_t.

+++ src/helper_functions.cpp
@@ -428,8 +428,8 @@ string HelperFunctions::parseString(const string& line, int str_pos, string& out
 	 */
 	bool HelperFunctions::matches(const string& s1, const string& s2)
 	{
-		unsigned int beg = 0, end;
-		unsigned int z;
+		size_t beg = 0, end;
+		size_t z;
 //		coutd << " s1 = " << s1 << ", s2 = " << s2 << endl;
 
 		//strings have matched to end - base case
